mac-capture: If window is not found, look for a matching one
If the target window does not exist, try to find one with a matching
title and application bundle ID. This way, I do not have to redo all my
window captures. This matches the behavior of the deprecated macOS
Window Capture source (except it still does not watch for the window to
be opened if it is closed when the capture starts).